rm xmlbuilder, move some stuff around
