+++ package.json
@@ -132,8 +132,7 @@
     "@types/node": "^12.12.11",
     "@types/sax": "^1.2.0",
     "arg": "^4.1.1",
-    "sax": "^1.2.4",
-    "xmlbuilder": "^13.0.2"
+    "sax": "^1.2.4"
   },
   "devDependencies": {
     "@babel/core": "^7.7.2",
